Add HexHook to allow "!pkg" to search hex.pm
When a command is issued in the format "!pkg SOME_TERM" the top 3
results will be displayed in the channel. One message per result.

To allow for the results to be displayed, a JSON decoder is required,
Poison has been added as a dependency to allow for this.

The description has been split into newlines and only the first is
displayed to avoid spamming the chat.